+++ packages/api/src/util/FeeExchange.ts
@@ -1,0 +1,15 @@
+import {AnyAssetId} from '@cennznet/types/types';
+
+export function generateTransactionPayment(tip: number, assetId: AnyAssetId, maxPayment: string) {
+  const feeExchange = {
+    assetId: assetId,
+    maxPayment: maxPayment,
+  };
+  const transactionPayment = {
+    tip: tip,
+    feeExchange: {
+      FeeExchangeV1: feeExchange,
+    },
+  };
+  return transactionPayment;
+}
